move code out of index.js routes
